fix & cleanup: move methods from gui.rs to struct & fix profile replacement on matching uuid's
